google_cloud/oauth_app/requirements.txt
Flask
google-api-python-client
google-auth-oauthlib
google-cloud-secret-manager